moved the utility file into a new folder called shared to allow for the update object to be used throughout the entire application
