+++ src/shared/utility.js
